test(lynx): tidy test runner split
