+++ examples/lynx-spa/package.json
@@ -8,8 +8,7 @@
     "dev": "rspeedy dev",
     "generate:icons": "node scripts/generate-icon-pages.js",
     "format": "biome format --write",
-    "preview": "rspeedy preview",
-    "test": "vitest run"
+    "preview": "rspeedy preview"
   },
   "dependencies": {
     "@karrotmarket/lynx-monochrome-icon": "1.10.0",
@@ -31,14 +30,11 @@
     "@lynx-js/rspeedy": "^0.13.6",
     "@lynx-js/types": "^3.7.0",
     "@rsbuild/plugin-type-check": "1.3.3",
-    "@testing-library/jest-dom": "^6.9.1",
     "@types/react": "^19",
-    "jsdom": "^27.4.0",
     "postcss": "^8.5.8",
     "tailwindcss": "^3.4.0",
     "sharp": "^0.34.5",
-    "typescript": "~5.9.3",
-    "vitest": "^3.2.4"
+    "typescript": "~5.9.3"
   },
   "engines": {
     "node": ">=18"
