+++ package.json
@@ -22,7 +22,9 @@
     "release": "bun packages:build && bun changeset publish",
     "version": "changeset version && bun install --no-immutable",
     "format:all": "bun biome format --fix",
-    "test:all": "bun test --path-ignore-patterns='packages/lynx-react/**' && bun lynx-react:test",
+    "test:all": "bun test:unit && bun test:lynx-react",
+    "test:unit": "bun test --path-ignore-patterns='packages/lynx-react/**'",
+    "test:lynx-react": "bun --filter @seed-design/lynx-react typecheck:test && bun --filter @seed-design/lynx-react test",
     "generate:all": "bun rootage:generate && bun qvism:generate && bun docs:generate",
     "figma:sync": "bun figma-extractor --config=scripts/.config/figma-extractor.config.ts scripts/data variables && bun run ./scripts/figma-to-rootage.ts",
     "images:convert": "bun scripts/convert-images-to-webp.ts",
@@ -49,8 +51,6 @@
     "headless:test:watch": "bun test --watch packages/react-headless",
     "react:test": "bun test packages/react",
     "react:test:watch": "bun test --watch packages/react",
-    "lynx-react:test": "bun --filter @seed-design/lynx-react test",
-    "lynx-react:test:watch": "bun --filter @seed-design/lynx-react test:watch",
     "clean:lib": "bun --filter '*' clean",
     "clean:modules": "find . -name 'node_modules' -type d -prune -exec rm -rf '{}' +",
     "generate:codemod-docs": "bun scripts/generate-codemod-docs.ts",
