+++ packages/lynx-qvism-preset/safe-area.test.ts
@@ -1,7 +1,7 @@
 import { expect, test } from "bun:test";
 
-import { globalCss } from "../src/global";
-import bottomSheet from "../src/recipes/bottom-sheet";
+import { globalCss } from "./src/global";
+import bottomSheet from "./src/recipes/bottom-sheet";
 
 test("globalCss does not define safe area variables with :root env fallback", () => {
   expect(globalCss[":root"]?.["--seed-safe-area-top"]).toBeUndefined();
