+++ packages/lynx-react/src/components/font-scale-contract.test.ts
